Command-defining macro(s)
Forward everything that's not :data by default.
Do not use alexandria:destructuring-case, because it warns on multiple
matches, which I want to have because I need default guards in the macro (i.e. to not forward data by default).

For now, `emit` is simply `(funcall forward :data data)`, if it needs
evolving in the future... we'll see.

fixes #4