Numbers!
Change parsing so that `2` is recognized as a number. Of course, this
will break when we want it explicitly as a string, but for now...
We'll allow `"2"` someday(tm).

Refactor parsing to prepare for collections.
Collections are arrays of tables with the same keys.
Tables are a singular object/dictionary, k=>v mapping.

The idea is that a collection will display itself as a properly
displayed table, SQL-like, so if you have two objects with key `a` and `b`, you will see:

```
| a | b |
|a1 |b1 |
|b1 |b2 |
```

But that's yet to be implemented!

Also (poorly) implement `single-number-arg`, but in the future the main
body of a command function will probably be a trivia match, extract its
arguments, and then call an underlying procedure.
Of course, even later down the line, we'll have a proper argparsing
facility. At that point, we might want to rename `calimero.parse` to
something else, but we'll see.